refactor: 공통되는 뷰 사용자 타입 뷰로 구현
